Move getSurveys to core
