+++ posthog-core/src/index.ts
@@ -23,6 +23,7 @@ export * as utils from './utils'
 import { LZString } from './lz-string'
 import { SimpleEventEmitter } from './eventemitter'
 import { uuidv7 } from './vendor/uuidv7'
+import { Survey, SurveyResponse } from './posthog-surveys-types'
 
 class PostHogFetchHttpError extends Error {
   name = 'PostHogFetchHttpError'
@@ -470,6 +471,29 @@ export abstract class PostHogCoreStateless {
       flags,
       payloads: parsedPayloads,
     }
+  }
+
+  /***
+   *** SURVEYS
+   ***/
+
+  public async getSurveys(): Promise<Survey[]> {
+    await this._initPromise
+
+    const url = `${this.host}/api/surveys/?token=${this.apiKey}`
+    const fetchOptions: PostHogFetchOptions = {
+      method: 'GET',
+      headers: this.getCustomHeaders(),
+    }
+
+    const response = await this.fetchWithRetry(url, fetchOptions)
+      .then((response) => response.json() as Promise<SurveyResponse>)
+      .catch((error) => {
+        this._events.emit('error', error)
+        return { surveys: [] }
+      })
+
+    return response.surveys
   }
 
   /***
