+++ posthog-react-native/src/surveys/components/ConfirmationMessage.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { StyleSheet, Text, View } from 'react-native'
 
 import { getContrastingTextColor, SurveyAppearanceTheme } from '../surveys-utils'
-import { SurveyQuestionDescriptionContentType } from '../posthog-surveys-types'
+import { SurveyQuestionDescriptionContentType } from '../../../../posthog-core/src/posthog-surveys-types'
 import { BottomSection } from './BottomSection'
 
 export function ConfirmationMessage({
