+++ posthog-react-native/src/surveys/components/QuestionTypes.tsx
@@ -15,7 +15,7 @@ import {
   LinkSurveyQuestion,
   MultipleSurveyQuestion,
   RatingSurveyQuestion,
-} from '../posthog-surveys-types'
+} from '../../../../posthog-core/src/posthog-surveys-types'
 import { BottomSection } from './BottomSection'
 import { QuestionHeader } from './QuestionHeader'
